+++ Realm/RLMNetworkClient.h
@@ -28,6 +28,7 @@ typedef NS_ENUM(NSUInteger, RLMServerEndpoint) {
     RLMServerEndpointLogout,
     RLMServerEndpointAddCredentials,
     RLMServerEndpointRemoveCredentials,
+    RLMServerEndpointChangePassword,
 };
 
 /**
@@ -42,11 +43,19 @@ NS_ASSUME_NONNULL_BEGIN
                          JSON:(NSDictionary *)jsonDictionary
                    completion:(RLMSyncCompletionBlock)completionBlock;
 
-/**
- Post some JSON data to the authentication server, and asynchronously call a completion block with a JSON response
- and/or error.
- */
 + (void)postRequestToEndpoint:(RLMServerEndpoint)endpoint
+                       server:(NSURL *)serverURL
+                         JSON:(NSDictionary *)jsonDictionary
+                      timeout:(NSTimeInterval)timeout
+                   completion:(RLMSyncCompletionBlock)completionBlock;
+
+/**
+ Send some JSON data using the specified HTTP method (e.g. 'POST', 'PUT', etc.) to the authentication server,
+ and asynchronously call a completion block with a JSON response and/or error.
+ */
+
++ (void)sendRequestToEndpoint:(RLMServerEndpoint)endpoint
+                   httpMethod:(NSString *)httpMethod
                        server:(NSURL *)serverURL
                          JSON:(NSDictionary *)jsonDictionary
                       timeout:(NSTimeInterval)timeout